runtime: split router.rs into the src/runtime/router/ module directory
Mechanical split of the 1099-line router.rs into mod.rs (module doc,
the route/router type aliases), sink.rs (BoundStarter, RouterSink),
routes.rs (the route structs, MountRoute, RouterDef), builder.rs (the
Router builder, mount tails, ComposedBlanket and the RouterDef/
MountRoute impls) and include.rs (the include family). Internal
re-exports keep every public path byte-identical; route-struct fields,
MountRoute and the mount tails became pub(super) so the sibling
submodules can reach them, and doc links that relied on the old super::
position now use full crate paths.

Stacked on the app.rs split: the old app.rs resolved [Router] links
through the router file module, which a directory module with a
re-export no longer satisfies; the app/ split already spells those
links in full.
